+++ python/plot3d/__init__.py
@@ -1,11 +1,12 @@
+from .block import Block, reduce_blocks
+from .blockfunctions import rotate_block,get_outer_bounds,block_connection_matrix
+from .connectivity import find_matching_blocks, get_face_intersection, connectivity_fast, face_matches_to_dict
 from .face import Face
-from .facefunctions import create_face_from_diagonals, split_face,outer_face_dict_to_list, match_faces_dict_to_list, find_connected_face, find_face
-from .block import Block
-from .blockfunctions import rotate_block, reduce_blocks
-from .connectivity import get_outer_faces, find_matching_blocks, get_face_intersection, connectivity, connectivity_fast, block_connection_matrix
+from .facefunctions import create_face_from_diagonals, get_outer_faces, find_connected_faces, find_bounding_faces,split_face,find_face_nearest_point,match_faces_dict_to_list,outer_face_dict_to_list,find_closest_block
 from .read import read_plot3D, read_ap_nasa
 from .write import write_plot3D
 from .differencing import find_edges, find_face_edges
-from .periodicity import periodicity, periodicity_fast, create_rotation_matrix, rotated_periodicity, translational_periodicity,translational_periodicity, translational_periodicity2
+from .periodicity import periodicity, periodicity_fast, create_rotation_matrix, rotated_periodicity, translational_periodicity
 from .point_match import point_match
 from .split_block import split_blocks, Direction
+from .listfunctions import unique_pairs
