+++ python/plot3d/blockfunctions.py
@@ -1,6 +1,12 @@
+from copy import deepcopy
+from itertools import combinations
+import math
 import numpy as np
-from typing import List
+from typing import Dict, List
+from tqdm import trange
+from .facefunctions import create_face_from_diagonals
 from .block import Block
+from .face import Face
 
 def rotate_block(block,rotation_matrix:np.ndarray) -> Block:
     """Rotates a block by a rotation matrix 
@@ -34,8 +40,6 @@ def rotate_block(block,rotation_matrix:np.ndarray) -> Block:
                 indx+=1
                 
     return Block(X,Y,Z)
-                    
-
 
 def reduce_blocks(blocks:List[Block],factor:int):
     """reduce the blocks by a factor of (factor)
@@ -97,3 +101,76 @@ def get_outer_bounds(blocks:List[Block]):
             zbounds[1] = zmax
     
     return tuple(xbounds),tuple(ybounds),tuple(zbounds)
+
+def block_connection_matrix(blocks:List[Block],outer_faces:List[Dict[str,int]]=[]):
+    """Creates a matrix representing how block edges are connected to each other 
+
+    Args:
+        blocks (List[Block]): _description_
+        outer_faces (List[Dict[str,int]], optional): List of outer faces remaining from connectivity. Useful if you are interested in finding faces that are exterior to the block. Also useful if you combine outerfaces with match faces, this will help identify connections by looking at split faces. Defaults to [].
+
+    Returns:
+        (Tuple): containing
+
+            *connectivity* (np.ndarray): integer matrix defining how the blocks are connected to each other
+            *connectivity_i* (np.ndarray): integer matrix defining connectivity of all blocks where IMAX=IMIN
+            *connectivity_j* (np.ndarray): integer matrix defining connectivity of all blocks where JMAX=JMIN
+            *connectivity_k* (np.ndarray): integer matrix defining connectivity of all blocks where KMAX=KMIN
+            
+    """
+    # Reduce the size of the blocks by the GCD 
+    gcd_array = list()    
+    for block_indx in range(len(blocks)):
+        block = blocks[block_indx]
+        gcd_array.append(math.gcd(block.IMAX-1, math.gcd(block.JMAX-1, block.KMAX-1)))
+    gcd_to_use = min(gcd_array) # You need to use the minimum gcd otherwise 1 block may not exactly match the next block. They all have to be scaled the same way.
+    blocks = reduce_blocks(deepcopy(blocks),gcd_to_use)
+
+    # Face to List 
+    outer_faces_all = list()
+    for o in outer_faces:
+        face = create_face_from_diagonals(blocks[o['block_index']], int(o['IMIN']/gcd_to_use), int(o['JMIN']/gcd_to_use), 
+            int(o['KMIN']/gcd_to_use), int(o['IMAX']/gcd_to_use), int(o['JMAX']/gcd_to_use), int(o['KMAX']/gcd_to_use))
+        face.set_block_index(o['block_index'])
+        outer_faces_all.append(face)
+
+    outer_faces = outer_faces_all
+
+    n = len(blocks)
+    connectivity = np.eye(n,dtype=np.int8)
+    combos = list(combinations(range(n),2))    
+    for indx in (pbar:=trange(len(combos))):
+        i,j = combos[indx]
+        pbar.set_description(f"Building block to block connectivity matrix: checking {i}")
+        b1 = blocks[i]
+
+        if len(outer_faces)==0:                     # Get the outerfaces to search
+            b1_outer_faces,_ = get_outer_faces(b1)
+        else:
+            b1_outer_faces = [o for o in outer_faces if o.BlockIndex == i]
+        
+        if i != j and connectivity[i,j]!=-1:
+            b2 = blocks[j]
+
+            if len(outer_faces)==0:                 # Get the outerfaces to search
+                b2_outer_faces,_ = get_outer_faces(b2)
+            else:
+                b2_outer_faces = [o for o in outer_faces if o.BlockIndex == j]                
+
+            # Check to see if any of the outer faces of the blocks match   
+            connection_found=False             
+            for f1 in b1_outer_faces:
+                for f2 in b2_outer_faces:
+                    if (f1.is_connected(f2)): # Check if face centroid is the same
+                        connectivity[i,j] = 1       # Default block to block connection matrix 
+                        connectivity[j,i] = 1
+                        connection_found=True
+                        break
+                if connection_found:
+                    break
+            if not connection_found:
+                connectivity[i,j] = -1
+                connectivity[j,i] = -1
+        # c = np.sum(connectivity[i,:]==1)
+        # print(f"block {i} connections {c}")
+    return connectivity
